Py_model/requirements.txt
pandas
joblib
numpy
paho-mqtt==1.6.1
scikit-learn